backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8.2,<2.9.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Orchestration
langchain==0.1.16
langgraph==0.0.40
langsmith>=0.1.0,<0.2.0
openai==1.6.1
tiktoken==0.5.2
tenacity==8.2.3

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database
weaviate-client==3.25.3
# pinecone-client==2.2.4  # Alternative to Weaviate

# Background Jobs
celery==5.3.4
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
clerk-backend-api==1.0.0
pyjwt>=2.9.0,<3.0.0
cryptography>=43.0.1,<44.0.0

# External APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
httpx>=0.27.0,<0.28.0
aiofiles==23.2.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring
prometheus-client==0.19.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
